task 101 corrections

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -11,7 +11,7 @@
 
 void print_number(int n)
 {
-	int div;
+	double div;
 
 	if (n < 0) /* print negative sign for negative numbers */
 	{
